Rename `PostEllipsisMenuContent` -> `InteractableEllipsisMenuContent`
